pyproject.toml
[tool.poetry]
name = "aironmangpt"
version = "0.1.0"
description = "AironmanGPT is an OSINT assistant based on DarkGPT (recommended use) designed to perform initial pentesting tasks. The idea is to use agents that encompass different pentesting tools, each one designed for a specific phase of pentesting."
authors = ["Alonso Isidoro Román <alonsoir@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
openai = "^1.34.0"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
pydantic = "^2.7.4"
black = "^24.4.2"
langchain = "^0.2.5"
langchain-openai = "^0.1.8"
loguru = "^0.7.2"
langchain-core = "^0.2.9"
sentencepiece = "^0.2.0"
protobuf = "^5.27.1"
premai = "^0.3.59"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
